Complete project re-structure splitting the classes, fixed the scoring logic
