+++ .gitignore
@@ -25,3 +25,8 @@ replay_pid*
 
 # Kotlin Gradle plugin data, see https://kotlinlang.org/docs/whatsnew20.html#new-directory-for-kotlin-data-in-gradle-projects
 .kotlin/
+
+# macos
+.DS_Store
+*.swp
+*.DS_Store
