--- .idea/modules/app/Dicing.app.iml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module version="4">
-  <component name="ModuleClassLoaderOverlays">
-    <paths>
-      <option value="/var/folders/kl/1_ll4g8x4ss7xg5z6qw73q8r0000gn/T/overlay1967323004350273136" />
-    </paths>
-  </component>
-</module>
