+++ app/src/main/java/com/sameerasw/dicing/DicingApp.kt
@@ -1,0 +1,8 @@
+package com.sameerasw.dicing
+
+import androidx.compose.runtime.Composable
+
+@Composable
+fun DicingApp() {
+    MainMenu()
+}
